backend/requirements.txt

Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
pymongo==4.8.0
python-dotenv==1.0.1
python-socketio==5.11.3
requests==2.32.3
twilio==9.2.3
gunicorn==20.1.0

